Refactor BroadcastCommand and remove deprecated BroadcastEventHandler
